src/project_name/requirements/qa.txt
# Todo: Define versions of all dependencies

# coverage measures code coverage of Python programs
# See http://coverage.readthedocs.io/en/latest/
coverage

pytest

# Todo: replace with Swagger
# **Sphinx** is used to create documentation
# See http://www.sphinx-doc.org/en/stable/tutorial.html
Sphinx

# **flake8** is used to check for PEP8 compliance
# See http://flake8.pycqa.org/en/latest/index.html
flake8

#core Django and DRF
Django
django-rest-framework